refactor: Decouple HsMeasureService
